cherry-pick https://github.com/THUDM/slime/pull/1070 "fix megatron cpu adam load issue".
This is an ad hoc patch for megatron. The origin issue is that megatron will assign step = None for the gpu optimizer in the hybrid optimizer which slime used as the pure cpu optimizer. In slime we won't use the gpu optimizer in most of the time, so this patch only fix the ckpt loading part.